refactor: move application converter interface to converter module
